fix: schema.py
