+++ core.py
@@ -1,13 +1,9 @@
-from os.path import isfile  # , isdir, makedirs
+from os.path import isfile
 
 from sqlalchemy import create_engine
 
 from models import (
     metadata_obj,
-    Nir_Grant,
-    Nir_NTP,
-    Nir_Templan,
-    Nir_VUZ,
     NTP,
     Grant,
     Templan,
@@ -17,10 +13,15 @@ from models import (
     create_pivot,
 )
 
+from schema import (
+    Nir_Grant,
+    Nir_NTP,
+    Nir_Templan,
+    Nir_VUZ,
+)
+
 
 def create_sql_tables():
-    # if not (isdir('DB')):
-    #     makedirs('DB')
     if not (isfile("DB/DataBase.sqlite")):
         engine = create_engine("sqlite:///DB/DataBase.sqlite", echo=False)
         metadata_obj.drop_all(engine)
@@ -30,3 +31,6 @@ def create_sql_tables():
         create_table(Nir_Templan, Templan)
         create_table(Nir_VUZ, VUZ)
         create_pivot(Grant, NTP, Templan, Pivot)
+        print('База данных DataBase.sqlite создана и подключена')
+    else:
+        print('База данных DataBase.sqlite подключена')
